initialize buffer in SML parser with 0 (fixes a bug int values that do not have the full length)

+++ nodes/ipt/gateway/src/controller.cpp
@@ -365,7 +365,7 @@ namespace node
 		//	random uint32
 		//
 		boost::random::mt19937 int_rng;
-        int_rng.seed(std::time(nullptr));
+        int_rng.seed(static_cast<std::uint32_t>(std::time(nullptr)));
 		boost::random::uniform_int_distribution<int> monitor_dist(10, 120);
 
 		//
